backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
mysqlclient
passlib==1.7.4
python-jose
python-dotenv
cryptography
bcrypt==3.2.0
matplotlib
numpy
mplcursors
plotly
jinja2
sqlalchemy-utils>=0.41.1
